Python API outline
Python API calls Rust extern "C" functions which call normal Rust
functions in the relational API
