+++ execution/src/relational_api/c_api.rs
@@ -1,0 +1,4 @@
+#[no_mangle]
+pub extern "C" fn test_interop() {
+    println!("It works!");
+}
